+++ packages/vendure-plugin-stripe-subscription/src/ui/queries.ts
@@ -44,3 +44,13 @@ export const DELETE_SCHEDULE = gql`
     deleteStripeSubscriptionSchedule(scheduleId: $scheduleId)
   }
 `;
+
+export const ELIGIBLE_PAYMENT_METHODS = gql`
+  query eligiblePaymentMethods {
+    eligiblePaymentMethods {
+      id
+      name
+      stripeSubscriptionPublishableKey
+    }
+  }
+`;